[billy-olsen, r=trival] Fix tox errors after namespace change.
Fix the tox errors resulting from the namespace change. The tox.ini file
still pointed to vmaas as the top level directory, but the new namespace
is maas_deployer. Additionally, the liberty was taken to refactor the
test case location to be at the root of the namespace and increase the
speed of the retry_on_exception test by mocking out the call to time.sleep.